pyproject.toml
[project]
name = "try-out"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.6",
    "jose>=1.0.0",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "setuptools>=75.8.0",
    "sqlalchemy>=2.0.37",
    "sqlmodel>=0.0.22",
]
